Upstream 4.4 change on src/s390

+++ src/s390/simulator-s390.h
@@ -384,7 +384,7 @@ class Simulator {
   // Simulator support.
   char* stack_;
   bool pc_modified_;
-  int icount_;
+  int64_t icount_;
 
   // Debugger input.
   char* last_debugger_input_;
